+++ adapters/coap-vertx/src/main/java/org/eclipse/hono/adapter/coap/impl/Config.java
@@ -14,8 +14,6 @@
 package org.eclipse.hono.adapter.coap.impl;
 
 import org.eclipse.hono.adapter.coap.CoapAdapterProperties;
-import org.eclipse.hono.client.RequestResponseClientConfigProperties;
-import org.eclipse.hono.config.ClientConfigProperties;
 import org.eclipse.hono.service.AbstractAdapterConfig;
 import org.eclipse.hono.service.metric.MetricsTags;
 import org.eclipse.hono.util.Constants;
@@ -60,41 +58,8 @@ public class Config extends AbstractAdapterConfig {
     }
 
     @Override
-    protected void customizeDownstreamSenderFactoryConfig(final ClientConfigProperties props) {
-        if (props.getName() == null) {
-            props.setName(CONTAINER_ID_HONO_COAP_ADAPTER);
-        }
-    }
-
-    @Override
-    protected void customizeRegistrationClientFactoryConfig(final RequestResponseClientConfigProperties props) {
-        if (props.getName() == null) {
-            props.setName(CONTAINER_ID_HONO_COAP_ADAPTER);
-        }
-    }
-
-    @Override
-    protected void customizeCredentialsClientFactoryConfig(final RequestResponseClientConfigProperties props) {
-        if (props.getName() == null) {
-            props.setName(CONTAINER_ID_HONO_COAP_ADAPTER);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void customizeTenantClientFactoryConfig(final RequestResponseClientConfigProperties props) {
-        if (props.getName() == null) {
-            props.setName(CONTAINER_ID_HONO_COAP_ADAPTER);
-        }
-    }
-
-    @Override
-    protected void customizeDeviceConnectionClientFactoryConfig(final RequestResponseClientConfigProperties props) {
-        if (props.getName() == null) {
-            props.setName(CONTAINER_ID_HONO_COAP_ADAPTER);
-        }
+    protected String getAdapterName() {
+        return CONTAINER_ID_HONO_COAP_ADAPTER;
     }
 
     /**